Support pattern modifiers to semantic module.
