+++ crates/semantic/src/expr/test_data/pattern
@@ -6,7 +6,7 @@ test_function_diagnostics
 //! > function
 
 func foo(x: (felt, felt)) {
-    let mut mut ref ref b = x;
+    let mut ref mut ref ref b = x;
 }
 
 //! > function_name
@@ -15,3 +15,17 @@ foo
 //! > module_code
 
 //! > expected_diagnostics
+error: `mut` modifier may not be repeated
+ --> lib.cairo:2:17
+    let mut ref mut ref ref b = x;
+                ^*^
+
+error: `ref` modifier may not be repeated
+ --> lib.cairo:2:21
+    let mut ref mut ref ref b = x;
+                    ^*^
+
+error: `ref` modifier may not be repeated
+ --> lib.cairo:2:25
+    let mut ref mut ref ref b = x;
+                        ^*^
