+++ crates/semantic/src/items/functions.rs
@@ -4,15 +4,13 @@ use db_utils::define_short_id;
 use debug::DebugWithDb;
 use defs::ids::{GenericFunctionId, GenericParamId, ParamLongId};
 use diagnostics_proc_macros::DebugWithDb;
-use syntax::node::ast::Modifier;
-use syntax::node::db::SyntaxGroup;
 use syntax::node::{ast, Terminal, TypedSyntaxNode};
 
 use crate::corelib::unit_ty;
 use crate::db::SemanticGroup;
-use crate::diagnostic::SemanticDiagnosticKind::RepeatedModifier;
 use crate::diagnostic::SemanticDiagnostics;
 use crate::expr::compute::Environment;
+use crate::items::modifiers::compute_modifiers;
 use crate::resolve_path::Resolver;
 use crate::types::{resolve_type, substitute_generics};
 use crate::{semantic, Parameter};
@@ -91,37 +89,6 @@ pub fn function_signature_return_type(
         }
     };
     resolve_type(db, diagnostics, resolver, &ty_syntax)
-}
-
-/// Returns the modifiers of a variable, given the list of modifiers in the AST.
-pub fn compute_modifiers(
-    diagnostics: &mut SemanticDiagnostics,
-    syntax_db: &dyn SyntaxGroup,
-    modifier_list: &[Modifier],
-) -> semantic::Modifiers {
-    let mut is_mut = false;
-    let mut is_ref = false;
-
-    for modifier in modifier_list {
-        match modifier {
-            Modifier::Mut(terminal) => {
-                if is_mut {
-                    diagnostics
-                        .report(terminal, RepeatedModifier { modifier: terminal.text(syntax_db) });
-                }
-                is_mut = true
-            }
-            Modifier::Ref(terminal) => {
-                if is_ref {
-                    diagnostics
-                        .report(terminal, RepeatedModifier { modifier: terminal.text(syntax_db) });
-                }
-                is_ref = true
-            }
-        }
-    }
-
-    semantic::Modifiers { is_mut, is_ref }
 }
 
 /// Returns the parameters of the given function signature's AST.
