+++ crates/semantic/src/items/mod.rs
@@ -4,5 +4,6 @@ pub mod extern_type;
 pub mod free_function;
 pub mod functions;
 pub mod generics;
+pub mod modifiers;
 pub mod strct;
 pub mod us;
